Rename md5 scenario to be platform neutral
gst_v4l2_video_decoder_md5_checksum_comparison becomes
gst_video_decoder_md5_checksum_comparison everywhere: the scenario is
about MD5 comparison, not about the v4l2 decoder family (Jetson uses
nvv4l2decoder, Renesas uses omx). Renames the scenario script and its
test module, the generator method, the template-id/filter/command, the
plan include, the genio-1200 conf key and the docs. Deliberate id
change for the genio-1200 md5 jobs (the only conf with the scenario);
the golden checksum directory under VIDEO_CODEC_TESTING_DATA follows
the scenario name, so the testing-data store layout must be renamed to
match. Also fixes the Vedio typo in the touched files.
