+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/tests/test_gst_video_decoder_md5_checksum_comparison.py
@@ -4,7 +4,7 @@ import subprocess
 import unittest
 from io import StringIO
 from unittest.mock import patch, MagicMock, mock_open, call
-from gst_v4l2_video_decoder_md5_checksum_comparison import (
+from gst_video_decoder_md5_checksum_comparison import (
     build_gst_command,
     extract_the_md5_checksum,
     get_md5_checksum_from_command,
@@ -97,7 +97,7 @@ class TestGetMD5ChecksumFromCommand(unittest.TestCase):
         logging.disable(logging.NOTSET)
 
     @patch(
-        "gst_v4l2_video_decoder_md5_checksum_comparison."
+        "gst_video_decoder_md5_checksum_comparison."
         "extract_the_md5_checksum"
     )
     @patch("subprocess.run")
@@ -163,10 +163,10 @@ class TestValidateVideoDecoderMD5Checksum(unittest.TestCase):
 
     @patch("os.path.exists")
     @patch(
-        "gst_v4l2_video_decoder_md5_checksum_comparison." "build_gst_command"
+        "gst_video_decoder_md5_checksum_comparison." "build_gst_command"
     )
     @patch(
-        "gst_v4l2_video_decoder_md5_checksum_comparison."
+        "gst_video_decoder_md5_checksum_comparison."
         "get_md5_checksum_from_command"
     )
     @patch(
@@ -210,7 +210,7 @@ class TestValidateVideoDecoderMD5Checksum(unittest.TestCase):
 
     @patch("os.path.exists")
     @patch(
-        "gst_v4l2_video_decoder_md5_checksum_comparison."
+        "gst_video_decoder_md5_checksum_comparison."
         "get_md5_checksum_from_command"
     )
     @patch("builtins.open", new_callable=mock_open)
@@ -249,7 +249,7 @@ class TestValidateVideoDecoderMD5Checksum(unittest.TestCase):
 
     @patch("os.path.exists")
     @patch(
-        "gst_v4l2_video_decoder_md5_checksum_comparison."
+        "gst_video_decoder_md5_checksum_comparison."
         "get_md5_checksum_from_command"
     )
     @patch("builtins.open", new_callable=mock_open)
@@ -291,10 +291,10 @@ class TestValidateVideoDecoderMD5Checksum(unittest.TestCase):
 
     @patch("os.path.exists")
     @patch(
-        "gst_v4l2_video_decoder_md5_checksum_comparison." "build_gst_command"
+        "gst_video_decoder_md5_checksum_comparison." "build_gst_command"
     )
     @patch(
-        "gst_v4l2_video_decoder_md5_checksum_comparison."
+        "gst_video_decoder_md5_checksum_comparison."
         "get_md5_checksum_from_command"
     )
     @patch(
