+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/units/video-codec/jobs.pxu
@@ -1,5 +1,5 @@
 id: video_codec_resource
-_summary: Generates mappings for all Vedio Codec Scenarios
+_summary: Generates mappings for all Video Codec Scenarios
 _description: Generate resource for all Video Codec scenarios.
 estimated_duration: 0.02
 category_id: video-codec
@@ -23,9 +23,9 @@ command:
 unit: template
 template-engine: jinja2
 template-resource: video_codec_resource
-template-filter: video_codec_resource.scenario == "gst_v4l2_video_decoder_md5_checksum_comparison"
+template-filter: video_codec_resource.scenario == "gst_video_decoder_md5_checksum_comparison"
 template-unit: job
-template-id: ce-oem-video-codec/gst_v4l2_video_decoder_md5_checksum_comparison
+template-id: ce-oem-video-codec/gst_video_decoder_md5_checksum_comparison
 _template-summary: To check if the MD5 checksum is same as golden reference under specific decoder
 id: ce-oem-video-codec/{{ name }}
 _summary: MD5 checksum comparison {{ width }}x{{ height }}-{{ decoder_plugin }}-{{ color_space }}
@@ -38,7 +38,7 @@ requires: manifest.has_video_codec == 'True'
 flags: also-after-suspend
 environ: GST_LAUNCH_BIN
 command:
-    gst_v4l2_video_decoder_md5_checksum_comparison.py -dp {{decoder_plugin}} -cs {{color_space}} -gp {{golden_sample_file}} -gmp {{golden_md5_checkum_file}}
+    gst_video_decoder_md5_checksum_comparison.py -dp {{decoder_plugin}} -cs {{color_space}} -gp {{golden_sample_file}} -gmp {{golden_md5_checkum_file}}
 
 unit: template
 template-engine: jinja2
