+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/units/video-codec/test-plan.pxu
@@ -23,7 +23,7 @@ _description: Automated tests for Video Codec in before suspend and post suspend
 bootstrap_include:
     video_codec_resource
 include:
-    ce-oem-video-codec/gst_v4l2_video_decoder_md5_checksum_comparison
+    ce-oem-video-codec/gst_video_decoder_md5_checksum_comparison
     ce-oem-video-codec/gst_video_decoder_performance_fakesink
     ce-oem-video-codec/gst_encoder_psnr
     ce-oem-video-codec/gst_transform_rotate_and_flip
